LU-5577 changelog: fix comparison between signed and unsigned
Change type of changelog_*{namelen,size}() to size_t.
Fixed string specifier for unsigned types.

Reviewed-on: http://review.whamcloud.com/12474
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/include/lustre/lustre_user.h
@@ -841,7 +841,7 @@ struct changelog_ext_rec {
 #define CHANGELOG_REC_EXTENDED(rec) \
 	(((rec)->cr_flags & CLF_VERMASK) == CLF_EXT_VERSION)
 
-static inline int changelog_rec_size(const struct changelog_rec *rec)
+static inline size_t changelog_rec_size(const struct changelog_rec *rec)
 {
 	return CHANGELOG_REC_EXTENDED(rec) ? sizeof(struct changelog_ext_rec):
 					     sizeof(*rec);
@@ -853,7 +853,7 @@ static inline const char *changelog_rec_name(const struct changelog_rec *rec)
 		((struct changelog_ext_rec *)rec)->cr_name: rec->cr_name;
 }
 
-static inline int changelog_rec_snamelen(const struct changelog_ext_rec *rec)
+static inline size_t changelog_rec_snamelen(const struct changelog_ext_rec *rec)
 {
 	return rec->cr_namelen - strlen(rec->cr_name) - 1;
 }
